rename: enums 폴더 분리 및 이동
