--- src/main/java/com/wanted/growthmate/payment/domain/PaymentMethod.java
@@ -1,5 +1,0 @@
-package com.wanted.growthmate.payment.domain;
-
-public enum PaymentMethod {
-    CARD
-}
