+++ src/main/java/com/wanted/growthmate/payment/domain/enums/PaymentMethod.java
@@ -1,0 +1,5 @@
+package com.wanted.growthmate.payment.domain.enums;
+
+public enum PaymentMethod {
+    CARD
+}
