+++ src/main/java/com/wanted/growthmate/payment/domain/enums/PaymentType.java
@@ -1,4 +1,4 @@
-package com.wanted.growthmate.payment.domain;
+package com.wanted.growthmate.payment.domain.enums;
 
 public enum PaymentType {
     // ---- 1차 구현 범위 ----
